Unifiy modules.
Only have a single module for the entire repository to make it simpler to
have code dependencies between the (formerly) different modules.

Modules in private repositories (like this one currently is) can cause
issues because they can't be easily fetched by builders without configuring
auth.  There's also skew, because tests for example will fetch the checked
in version, not the version in the repository.

We'd attempted to work around this with replace directive in go.mod which
was great for local development, but didn't help other cases.