pipeline: make /publish require POST
It looks like the intent of the code is to require the POST method for
all calls to /publish. However, when the http.HandlerFunc discovers that
the current call is using the wrong method, it only sets the HTTP status
code and does not actually stop processing the request.

To prevent acting on non-POST calls, we should bail out immediately
after setting that status code.