pipeline: ensure file is on disk before upload
Some filesystems can delay the reporting of write errors until the
program closes the file. Use an ioutil helper to be sure we detect any
problems before attempting an upload.

On top of that, using the helper ensures that the data is on disk before
executing another program to do the upload. I don't know all the rules
of when writes are visible to other processes; it seems like this could
be a problem, and could combine with the previous use of bufio.Writer
(which does not see a call to Flush) in a confusing way.

Finally, use path/filepath when manipulating paths on disk. It shouldn't
matter either way.