airtable: Switch to pure-go download implementation.
Shelling out to Python, which reads the JSON and serializes it to
disk, so we can read and parse that, is slow -- taking ~55s to
complete.  Some of this latency is unavoidable; Airtable results are
returned in pages of 100, and are limited to 5 req/s, making 12s the
_minimum_ possible time for downloading the Locations table at its
current size.  A significant portion of the overhead is in parsing and
serializing to disk the JSON.

Switch to a Go implementation of `airtable-export`, which is in
actuality rather simple[1].  This also allows for potentially better
insight into errors.

With this change, the time (over 10 runs) to run `/publish`
dropped from (mean ± σ) of 63s ± 11s, to 36s ± 2.9s.

[1] https://github.com/simonw/airtable-export/blob/main/airtable_export/cli.py#L74-L90