airtable: Also cache errors, not just successes.
If a request times out or otherwise failed, we previously returned the
`err` and left the table uncached; this meant any parallel fetches
which were blocked would effectively retry it.

If the failures were _slow_ failures, and persistent, this would
result in an overall extremely slow failures.  Instead, cache and
return the same err for all other fetches, to better bound the
response time, and make the results consistent for the resource across
versions.