storage: Provide a `-local` storage that does not need Google Cloud.
When passed `-local`, write out the files to a local directory,
instead of attempting to upload to Google Cloud.  With this option,
having local Google Cloud credentials is no longer necessary to test;
`-metrics=false` (the default on `once`) and `-local` mean that the
only necessary secret is the `AIRTABLE_KEY`.