freshcf: Switch to secrets.Get, from os.Getenv.
d40494609cbcab7d0d1104d8ad0a2f3b547acf9e changed the metrics package
in pipeline, but not in freshcf; it removed the `HONEYCOMB_KEY`
environment variable, leaving it empty for the old code that went
looking for it in freshcf.

Update to use `secrets.Get`.  Unfortunately, we cannot factor the
`beeline.Config` value out into a helper in the `config` package, like
was done for `StackdriverOptions`, because it references the `deploy`,
which can't be imported into `config` without causing a circular
dependency.